Fixing analyzer violation
S3928: The parameter name 'OrchardCoreConfiguration should be provided.' is not declared in the argument list.
  C:\Agent2\work\FINI\test\Lombiq.UITestingToolbox\Lombiq.Tests.UI\Services\UITestExecutor.cs(488,23):
CA2208: Method ExecuteOrchardCoreTestAsync passes 'OrchardCoreConfiguration should be provided.' as the paramName argument to a ArgumentNullException constructor. Replace this argument with one of the method's parameter names. Note that the provided parameter name should have the exact casing as declared on the method.