Remove FakeStore
The service seems unnecessary. It's only purpose was to keep track of sessions and dispose them
explicitly when disposing OrchardApplicationFactory. However, this happens automatically when
disposing the host anyway.

It also was something to maintain with YesSql upgrades, and with the OC 2.2.0 update causing
and incompatibility due to YesSql breaking changes (which added new methods to IStore).
Fixing that would need referencing the latest YesSql directly in this project, making it incompatible
with OC <2.2.0.